+++ src/main/kotlin/gonorth/console/Input.kt
@@ -1,7 +1,6 @@
 package gonorth.console
 
 import arrow.core.None
-import arrow.core.getOrElse
 import arrow.core.toOption
 import gonorth.GoNorth
 import gonorth.domain.*
@@ -18,7 +17,7 @@ interface SimpleClient {
 class ConsoleClient(private val engine: GoNorth,
                     private val console: Console,
                     private val worldBuilder: GameStateGenerator,
-                    private val parser: PossibilityGenerator) : SimpleClient {
+                    private val parser: PossibilityFilter) : SimpleClient {
 
     override fun startGame(seed: Long): GameState {
         val r = Random(seed).nextLong()
@@ -41,7 +40,7 @@ class ConsoleClient(private val engine: GoNorth,
         }
 
         fun handleInputs(gameState: GameState, console: Console): GameState {
-            val inputChoices = parser.generate(gameState)
+            val inputChoices = parser.filter(gameState)
             val topLevelChoices = rootChoices(inputChoices)
 
             if (topLevelChoices.isNotEmpty())
@@ -95,7 +94,7 @@ class ConsoleClient(private val engine: GoNorth,
         return handleInputs(gameState, console)
     }
 
-    fun <K, A, B> Map<K, A>.foldLeft(b: B, f: (B, Map.Entry<K, A>) -> B): B {
+    private fun <K, A, B> Map<K, A>.foldLeft(b: B, f: (B, Map.Entry<K, A>) -> B): B {
         var result = b
         this.forEach { result = f(result, it) }
         return result
@@ -105,55 +104,3 @@ class ConsoleClient(private val engine: GoNorth,
 typealias Choices = Map<Char, String>
 
 data class InputChoices(val movement: Choices, val describe: Choices, val take: Choices, val use: Choices)
-
-class PossibilityGenerator {
-
-    fun generate(gameState: GameState): InputChoices {
-
-        val emptySelections = emptyMap<Char, String>()
-
-        val movement = gameState
-                .fetchLinks(gameState.currentLocation)
-                .getOrElse { emptySet() }
-                .fold(emptySelections, { m, l ->
-                    when (l.move) {
-                        Move.NORTH -> m.plus(Pair('w', "North"))
-                        Move.EAST -> m.plus(Pair('d', "East"))
-                        Move.SOUTH -> m.plus(Pair('s', "South"))
-                        Move.WEST -> m.plus(Pair('a', "West"))
-                        else -> m.plus(Pair('x', "???"))
-                    }
-                })
-
-        val describe = gameState.locationOpt()
-                .map { it.items }
-                .getOrElse { emptySet() }
-                .toSortedSet(java.util.Comparator { o1, o2 -> o1.name().compareTo(o2.name()) })
-                .fold(emptySelections, { m, l ->
-                    m.plus(((1 + m.size).toString().first()) to l.name())
-                })
-
-        val take = gameState.locationOpt()
-                .map { it.items.onlyItems() }
-                .getOrElse { emptyList() }
-                .fold(emptySelections, { m, l ->
-                    m.plus(((1 + m.size).toString().first()) to l.name())
-                })
-
-
-        val fixedUse = gameState.locationOpt()
-                .map { it.items.onlyFixed() }
-                .getOrElse { emptyList() }
-                .fold(emptySelections, { m, l ->
-                    m.plus(((1 + m.size).toString().first()) to l.name())
-                })
-
-        val use = gameState.player.inventory
-                .filter { it.requiredLocation.map { it == gameState.currentLocation }.getOrElse { true } }
-                .fold(fixedUse, { m, l ->
-                    m.plus(((1 + m.size).toString().first()) to l.name())
-                })
-
-        return InputChoices(movement, describe, take, use)
-    }
-}
